fix: 迁移 normalized 测试到 im_adapter/normalized.rs
将 src/im/normalized_tests.rs 中的 3 个单元测试迁移到
src/im_adapter/normalized.rs，使测试与 NormalizedMessage
类型定义同文件，遵循 Rust 模块测试惯例。

- 删除 src/im/normalized_tests.rs
- 从 src/im/mod.rs 移除 normalized_tests 模块声明
- use 路径改为 super::NormalizedMessage

Source: issue #901
Type: fix
